+++ HISTORY.rst
@@ -3,6 +3,14 @@
 Release History
 ---------------
 
+1.10.6 (2020-08-10)
++++++++++++++++++++
+
+**Bug Fixes**
+
+- Prevent closed markets being removed when paper trading
+- Fix missing MarketBook from market (closes #FLUMINE-PROD-EE)
+
 1.10.5 (2020-08-04)
 +++++++++++++++++++